Extract temporary account modification to build tag protected file
