+++ pkg/acceptance/helpers/parameter_client_account_level.go
@@ -1,0 +1,32 @@
+//go:build account_level_tests
+
+package helpers
+
+import (
+	"context"
+	"testing"
+
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
+	"github.com/stretchr/testify/require"
+)
+
+func (c *ParameterClient) UpdateAccountParameterTemporarily(t *testing.T, parameter sdk.AccountParameter, newValue string) func() {
+	t.Helper()
+	ctx := context.Background()
+
+	param := c.ShowAccountParameter(t, parameter)
+	oldValue := param.Value
+	oldLevel := param.Level
+
+	err := c.client().SetAccountParameter(ctx, parameter, newValue)
+	require.NoError(t, err)
+
+	return func() {
+		if oldLevel == "" {
+			c.UnsetAccountParameter(t, parameter)
+		} else {
+			err := c.client().SetAccountParameter(ctx, parameter, oldValue)
+			require.NoError(t, err)
+		}
+	}
+}
