+++ pkg/sdk/testint/security_integrations_gen_integration_test.go
@@ -1,3 +1,5 @@
+//go:build account_level_tests
+
 package testint
 
 import (
